requirements.txt

base58==2.1.1
blinker==1.4
blocksmith==1.2.1
cachetools==5.2.0
certifi==2022.5.18.1
cfgv==3.3.1
click==8.1.3
cloudpickle==2.1.0
colorful==0.5.4
colour==0.1.5
cssselect2==0.6.0
cssutils==1.0.2
dash==2.5.1
dash-bootstrap-components==1.2.0
dash-core-components==2.0.0
dash-dangerously-set-inner-html==0.0.2
dash-html-components==2.0.0
dash-auth==1.3.2
dash-table==5.0.0
dash-trich-components==1.0.0
filelock==3.7.1
Flask==2.1.2
Flask-Compress==1.12
Flask-Cors==3.0.10
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.1
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
html5lib==1.1
identify==2.5.1
importlib-metadata==4.11.4
iniconfig==1.1.1
gunicorn==20.0.4
ipython-genutils==0.2.0
ipywidgets==7.7.1
iso8601==1.0.2
joblib==1.1.0
JPype1==1.4.0
json2html==1.3.0
waitress==2.1.2
libclang==14.0.1
lxml==4.9.0
Markdown==3.3.7
msgpack==1.0.3
multidict==6.0.2
multitasking==0.0.10

nodeenv==1.6.0
oauthlib==3.2.0
opencensus==0.9.0
opencensus-context==0.1.2
opt-einsum==3.3.0
oscrypto==1.3.0
outcome==1.2.0
pandas==1.4.2
pandas-datareader==0.10.0
pandocfilters==1.4.2
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11

protobuf==3.19.4
psutil==5.9.1
py==1.11.0
py-spy==0.3.12
pycryptodome==3.15.0
pydeck==0.7.1
pyhanko-certvalidator==0.19.5
Pympler==1.0.1
pyparsing==2.4.7
pytrends==4.8.0
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.11
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
rich==12.5.1
rsa==4.8
selenium==4.4.3
semver==2.13.0
six==1.15.0
smart-open==6.0.0
smmap==5.0.0
sortedcontainers==2.4.0
svglib==1.3.0
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
threadpoolctl==3.1.0
thriftpy2==0.4.14
tinycss2==1.1.1
trio==0.21.0
trio-websocket==0.9.2

tzlocal==4.2
uritools==4.0.0
webencodings==0.5.1
websockets==10.3
xhtml2pdf==0.2.8


